Adding Required Classes
This commit involves adding all required classes that have been
specified in the class diagram in the prelim. spec. A few of the classes
like game, player, and ai have basic outlines for the methods that would
be involved and Location has a good amount of work done already, so this
class is close to done. Note that an additional class, "TestingInCode"
has been added so that I can make sure everything works fine when
running/before commiting